setup config-server and config-client

+++ app-config-data/src/main/java/com/microservices/demo/config/RetryConfigData.java
@@ -11,7 +11,7 @@ import org.springframework.context.annotation.Configuration;
 public class RetryConfigData {
     private Long initialIntervalMs;
     private Long maxIntervalMs;
-    private Integer multiplier;
+    private Double multiplier;
     private Integer maxAttempts;
     private Long sleepTimeMs;
 }
